requirements.txt
async-timeout==4.0.2
mysql-connector-python==8.0.33
pip==22.3.1
protobuf==3.20.3
redis==4.5.4
setuptools==65.5.1
wheel==0.38.4
